refactor: Make DegreePlannerChrome more usable
This is the big kahuna that should not have taken months to finish but
did anyway. The DegreePlannerChrome now uses hooks to abstract away most
core functionality, including handling drag and drop, undo/redo (to be
implemented), UX flows for performing CRUD operations for individual
semesters and a CoursePlan as a whole, and more. If the commit author
were of legal age, it may be appropriate to say he deserves a nice drink
after this behemmoth of a rewrite.

TODO: Re-enable drag and drop