feat(planner): Reimplement drag-and-drop interface and planner UI
There is now a DraggableItemContainer that exposes a high-level API for manipulating draggable lists
of items. More work is needed to finalize the API surface, like adding proper support for additional
children rendered in the list. Additionally, the /app/plans/[...planId] route now supports a whole
suite of features. These may not make it to the final release as they haven't been implemented yet
aside from the core planner UI, but they include an overview of one's degree plan status, attempted
course history, and degree plan requirements.

BREAKING CHANGE: The DegreePlannerChrome is no longer used for drag-and-drop functionality and is
deprecated in favor of the DraggableItemContainer in the [...planId].tsx file.