feat(planner): Allow plan import from file
The planner UI now has a "Import plan" button that allows a user to select a
JSON file containing plan contents and import it into the planner. Importing a
plan replaces the currently loaded plan in memory.

TODO: Fix issue with input field not firing event for onChange when selecting
a plan with the same file name as the last imported plan
TODO: Automatically save imported plan to local storage
Resolve #48