feat(planner): Create ability to add courses to planner
After courses are selected and the button "Add Courses" is pushed inside
components/planner/CourseSelector.tsx, a Droppable is created inside
components/planner/DraggableItemContainer.tsx for the user to drag courses into
the planner. This commit has created the core logic to add courses. Styling of
the components & adding further non-essential logic sitll needs to be done and
will be addressed in future commits.