feat(planner): Create dialog for generating a new degree plan
When a user is in "app/plans", there is now a button that opens a dialog for the
user to create a new degree plan. Once the user is finished filling out the
fields, the dialog will close and the user will be redirected to
"app/plans/[planId]". Most of the functionality in NewPlanDialog.tsx still needs
to be implemented, as well as generating the degree plan based on user input.